added issue nft test, decimals and assetName boundary tests

+++ node-it/src/test/scala/com/wavesplatform/it/api/AsyncHttpApi.scala
@@ -783,7 +783,7 @@ object AsyncHttpApi extends Assertions {
     def broadcastIssue(source: KeyPair,
                        name: String,
                        quantity: Long,
-                       decimals: Byte,
+                       decimals: Int,
                        reissuable: Boolean,
                        fee: Long,
                        description: ByteString = ByteString.EMPTY,
